# AVX-2 accelerated SAD functions to speedup motion estimation