Here are the obtained bench results with a unaligned accesses:
```
Benchmarking get_sad/BLOCK_4X4: Collecting 100 samples in estimated 5.0000 s (28                                                                                get_sad/BLOCK_4X4       time:   [12.919 ns 12.926 ns 12.935 ns]
                        change: [-7.8931% -7.7870% -7.6862%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
Benchmarking get_sad/BLOCK_4X8: Collecting 100 samples in estimated 5.0000 s (21                                                                                get_sad/BLOCK_4X8       time:   [19.543 ns 19.551 ns 19.561 ns]
                        change: [-6.0339% -5.9476% -5.8651%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
Benchmarking get_sad/BLOCK_8X4: Collecting 100 samples in estimated 5.0000 s (22                                                                                get_sad/BLOCK_8X4       time:   [18.048 ns 18.063 ns 18.082 ns]
                        change: [-5.5344% -5.4019% -5.2604%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low severe
  5 (5.00%) high mild
  8 (8.00%) high severe
Benchmarking get_sad/BLOCK_8X8: Collecting 100 samples in estimated 5.0001 s (27                                                                                get_sad/BLOCK_8X8       time:   [13.736 ns 13.742 ns 13.747 ns]
                        change: [-6.8187% -6.7357% -6.6548%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_8X16: Collecting 100 samples in estimated 5.0001 s (1                                                                                get_sad/BLOCK_8X16      time:   [22.036 ns 22.047 ns 22.059 ns]
                        change: [-2.7995% -2.6955% -2.5958%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) high mild
  6 (6.00%) high severe
Benchmarking get_sad/BLOCK_16X8: Collecting 100 samples in estimated 5.0001 s (2                                                                                get_sad/BLOCK_16X8      time:   [20.582 ns 20.604 ns 20.629 ns]
                        change: [-3.0224% -2.7958% -2.6272%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_16X16: Collecting 100 samples in estimated 5.0001 s (                                                                                get_sad/BLOCK_16X16     time:   [17.758 ns 17.768 ns 17.778 ns]
                        change: [-51.338% -51.294% -51.248%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_16X32: Collecting 100 samples in estimated 5.0001 s (                                                                                get_sad/BLOCK_16X32     time:   [29.969 ns 29.987 ns 30.005 ns]
                        change: [-54.651% -54.593% -54.540%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) low severe
  4 (4.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_32X16: Collecting 100 samples in estimated 5.0000 s (                                                                                get_sad/BLOCK_32X16     time:   [29.423 ns 29.441 ns 29.461 ns]
                        change: [-54.971% -54.927% -54.884%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  6 (6.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_32X32: Collecting 100 samples in estimated 5.0000 s (                                                                                get_sad/BLOCK_32X32     time:   [25.039 ns 25.069 ns 25.098 ns]
                        change: [-79.765% -79.739% -79.716%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 22 outliers among 100 measurements (22.00%)
  8 (8.00%) low severe
  3 (3.00%) low mild
  7 (7.00%) high mild
  4 (4.00%) high severe
Benchmarking get_sad/BLOCK_32X64: Collecting 100 samples in estimated 5.0002 s (                                                                                get_sad/BLOCK_32X64     time:   [45.083 ns 45.121 ns 45.157 ns]
                        change: [-81.788% -81.762% -81.733%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) low severe
  4 (4.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe
Benchmarking get_sad/BLOCK_64X32: Collecting 100 samples in estimated 5.0002 s (                                                                                get_sad/BLOCK_64X32     time:   [43.684 ns 43.728 ns 43.773 ns]
                        change: [-82.089% -82.060% -82.033%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  8 (8.00%) low severe
  4 (4.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe
Benchmarking get_sad/BLOCK_64X64: Collecting 100 samples in estimated 5.0004 s (                                                                                get_sad/BLOCK_64X64     time:   [62.709 ns 62.743 ns 62.782 ns]
                        change: [-86.687% -86.669% -86.651%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_64X128: Collecting 100 samples in estimated 5.0001 s                                                                                 get_sad/BLOCK_64X128    time:   [166.02 ns 169.76 ns 173.26 ns]
                        change: [-83.166% -82.930% -82.702%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 22 outliers among 100 measurements (22.00%)
  22 (22.00%) high severe
Benchmarking get_sad/BLOCK_128X64: Collecting 100 samples in estimated 5.0006 s                                                                                 get_sad/BLOCK_128X64    time:   [119.19 ns 119.35 ns 119.54 ns]
                        change: [-87.075% -87.055% -87.034%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe
Benchmarking get_sad/BLOCK_128X128: Collecting 100 samples in estimated 5.0030 s                                                                                get_sad/BLOCK_128X128   time:   [523.51 ns 524.07 ns 524.70 ns]
                        change: [-74.082% -74.040% -73.998%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
Benchmarking get_sad/BLOCK_4X16: Collecting 100 samples in estimated 5.0000 s (1                                                                                get_sad/BLOCK_4X16      time:   [32.774 ns 32.787 ns 32.801 ns]
                        change: [-3.7121% -3.6143% -3.5287%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_16X4: Collecting 100 samples in estimated 5.0000 s (1                                                                                get_sad/BLOCK_16X4      time:   [28.848 ns 28.881 ns 28.929 ns]
                        change: [-2.8088% -2.6928% -2.5612%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low mild
  8 (8.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_8X32: Collecting 100 samples in estimated 5.0002 s (1                                                                                get_sad/BLOCK_8X32      time:   [38.372 ns 38.385 ns 38.399 ns]
                        change: [-1.5874% -1.4598% -1.3111%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe
Benchmarking get_sad/BLOCK_32X8: Collecting 100 samples in estimated 5.0002 s (1                                                                                get_sad/BLOCK_32X8      time:   [35.230 ns 35.265 ns 35.289 ns]
                        change: [-1.9183% -1.7353% -1.5565%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 21 outliers among 100 measurements (21.00%)
  9 (9.00%) low severe
  9 (9.00%) high mild
  3 (3.00%) high severe
Benchmarking get_sad/BLOCK_16X64: Collecting 100 samples in estimated 5.0003 s (                                                                                get_sad/BLOCK_16X64     time:   [54.226 ns 54.266 ns 54.309 ns]
                        change: [-56.833% -56.792% -56.752%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  8 (8.00%) low mild
  4 (4.00%) high mild
  6 (6.00%) high severe
Benchmarking get_sad/BLOCK_64X16: Collecting 100 samples in estimated 5.0002 s (                                                                                get_sad/BLOCK_64X16     time:   [52.418 ns 52.458 ns 52.506 ns]
                        change: [-56.556% -56.511% -56.462%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  8 (8.00%) low mild
  5 (5.00%) high mild
  6 (6.00%) high severe
```