@barrbrain 2019-04-02
This is great and the changes to benches/me.rs are no longer needed. (Can be dropped at rebase.)
Just for the record, these are the numbers @tdaede got when benchmarking with  a baseline of #1143:

```
get_sad/BLOCK_4X4       time:   [15.453 ns 15.454 ns 15.455 ns]                               
                        change: [-31.022% -31.008% -30.996%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low severe
  2 (2.00%) low mild
  6 (6.00%) high mild
  1 (1.00%) high severe
get_sad/BLOCK_4X8       time:   [30.076 ns 30.081 ns 30.085 ns]                              
                        change: [-2.7481% -2.7182% -2.6928%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
get_sad/BLOCK_8X4       time:   [27.425 ns 27.428 ns 27.430 ns]                               
                        change: [-3.4026% -3.3732% -3.3428%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe
get_sad/BLOCK_8X8       time:   [17.176 ns 17.177 ns 17.178 ns]                               
                        change: [-27.664% -27.656% -27.650%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low severe
  3 (3.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe
get_sad/BLOCK_8X16      time:   [33.520 ns 33.522 ns 33.525 ns]                               
                        change: [-2.4336% -2.4082% -2.3660%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
get_sad/BLOCK_16X8      time:   [30.476 ns 30.478 ns 30.480 ns]                               
                        change: [-3.8034% -3.7931% -3.7816%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe
get_sad/BLOCK_16X16     time:   [22.126 ns 22.141 ns 22.156 ns]                                 
                        change: [-55.375% -55.326% -55.268%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
get_sad/BLOCK_16X32     time:   [44.531 ns 44.538 ns 44.546 ns]                                
                        change: [-48.014% -47.978% -47.930%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  6 (6.00%) high mild
  8 (8.00%) high severe
get_sad/BLOCK_32X16     time:   [41.643 ns 41.651 ns 41.661 ns]                                
                        change: [-50.819% -50.788% -50.744%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe
get_sad/BLOCK_32X32     time:   [30.111 ns 30.114 ns 30.117 ns]                                
                        change: [-80.300% -80.282% -80.257%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  5 (5.00%) high mild
  9 (9.00%) high severe
get_sad/BLOCK_32X64     time:   [61.459 ns 61.475 ns 61.493 ns]                                
                        change: [-79.844% -79.827% -79.804%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe
get_sad/BLOCK_64X32     time:   [71.771 ns 71.776 ns 71.781 ns]                                
                        change: [-75.796% -75.776% -75.751%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  5 (5.00%) high mild
  9 (9.00%) high severe
get_sad/BLOCK_64X64     time:   [131.28 ns 131.30 ns 131.33 ns]                                
                        change: [-76.749% -76.732% -76.709%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
get_sad/BLOCK_64X128    time:   [317.02 ns 317.10 ns 317.19 ns]                                  
                        change: [-74.314% -74.207% -74.090%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) low mild
  6 (6.00%) high mild
  5 (5.00%) high severe
get_sad/BLOCK_128X64    time:   [263.03 ns 263.18 ns 263.33 ns]                                 
                        change: [-76.003% -75.984% -75.960%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
get_sad/BLOCK_128X128   time:   [789.06 ns 789.60 ns 790.14 ns]                                   
                        change: [-68.450% -68.427% -68.402%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
get_sad/BLOCK_4X16      time:   [47.717 ns 47.719 ns 47.720 ns]                               
                        change: [+0.4193% +0.4370% +0.4553%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe
get_sad/BLOCK_16X4      time:   [39.040 ns 39.046 ns 39.054 ns]                               
                        change: [-2.2906% -2.2457% -2.1981%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low severe
  2 (2.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe
get_sad/BLOCK_8X32      time:   [54.730 ns 54.737 ns 54.747 ns]                               
                        change: [-0.1244% -0.0966% -0.0705%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low severe
  4 (4.00%) high mild
  7 (7.00%) high severe
get_sad/BLOCK_32X8      time:   [45.214 ns 45.219 ns 45.224 ns]                               
                        change: [-6.9701% -6.9476% -6.9202%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe
get_sad/BLOCK_16X64     time:   [75.599 ns 75.604 ns 75.610 ns]                                
                        change: [-52.299% -52.269% -52.227%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe
get_sad/BLOCK_64X16     time:   [75.092 ns 75.103 ns 75.114 ns]                                
                        change: [-49.672% -49.636% -49.591%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  8 (8.00%) high mild
  6 (6.00%) high severe
```